requirements.txt
# Core
streamlit>=1.36.0
python-dotenv>=1.0.1
pydantic>=2.7.0

# LLM
anthropic>=0.39.0

# RAG / orchestration
langchain>=0.3.0
langchain-community>=0.3.0
langchain-anthropic>=0.2.0
langchain-text-splitters>=0.3.0

# Vector DB
chromadb>=0.5.5

# Embeddings (PubMedBERT via sentence-transformers)
sentence-transformers>=3.0.1
torch>=2.2.0

# Document loaders
pypdf>=4.2.0
python-docx>=1.1.2
beautifulsoup4>=4.12.3

# Utilities
tqdm>=4.66.4
tenacity>=8.5.0
